+++ documentation/docs/src/lib/components/FeedBack/Modals/ModalManager/ModalManager.stories.svelte
@@ -62,6 +62,27 @@
   <ModalManager />
 </Story>
 
+<Story name="Confirm Actions Demo" asChild>
+  <Button
+    onclick={() => {
+      modals.add('confirm', {
+        title: 'Confirm Actions Demo',
+        description: 'Demo for testing actions',
+        onaccept: () => {
+          alert('onaccept clicked');
+        },
+        onreject: () => {
+          alert('onreject clicked');
+        },
+      });
+    }}
+  >
+    Add Modal
+  </Button>
+
+  <ModalManager />
+</Story>
+
 <Story name="Modal master with custom ID" asChild>
   <Button
     onclick={() => {