+++ documentation/customize/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @documentation/customize
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @flightlesslabs/dodo-ui-bits@0.10.11
+
 ## 0.0.4
 
 ### Patch Changes