+++ packages/dodo-ui-bits/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flightlesslabs/dodo-ui-bits
 
+## 0.10.11
+
+### Patch Changes
+
+- Modal manager accept/reject action fixed
+
 ## 0.10.10
 
 ### Patch Changes